config/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
django-celery-beat==2.4.0
django-stubs==1.13.1
django-stubs-ext==0.7.0
django-timezone-field==5.0
djangorestframework==3.14.0
idna==3.4
kombu==5.2.4
mypy==0.991
mypy-extensions==0.4.3
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
python-crontab==2.7.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
